TravisCI: Use more portable module ver stripping.
This modifies the script that strips the module version from listed
modules for the purposes of changing to the relevant directory to be
more portable by adding the -E flag to sed.  This allows it to work
properly on Mac as well.

While here, it also switches the sed separator to a comma in the
relevant removals to avoid the need to escape the slashes in the path.