release: Freeze version 2 rpcclient module.
This serves as the final release of version 2 of the rpcclient module.
All future releases will be moving to version 3 of the module.

Consequently, it bumps the module version as follows:

- github.com/decred/dcrd/rpcclient@v2.1.0

In addition, it updates the main module to use following module versions
as required:

- github.com/decred/dcrd/dcrutil@v1.4.0

It also removes the rpcclient override in the root module and updates it
so building the software will still produce binaries based on the v2
module until the v3 module is fully released.