hdkeychain: Prepare v2.0.1.
This updates the hdkeychain dependencies and serves as a base for
hdkeychain/v2.0.1.

Note that even though some of the updated dependencies consume a new
major version, they are only used internally or in the example code and
therefore a major version bump of the hdkeychain module is not required.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/v2@v2.0.2
- github.com/decred/dcrd/dcrutil/v2@2.0.0
- github.com/decred/dcrd/dcrec/secp256k1@v1.0.2

The full list of updated direct dependencies since the previous
hdkeychain/v2.0.0 release are as follows:

- github.com/decred/dcrd/chaincfg/v2@v2.0.2
- github.com/decred/dcrd/dcrec@v1.0.0
- github.com/decred/dcrd/dcrutil/v2@2.0.0
- github.com/decred/dcrd/dcrec/secp256k1@v1.0.2