release: Freeze version 1 database module.
This serves as the final release of version 1 of the database module.
All future releases will be moving to version 2 of the module.

Consequently, it removes the database override in the root module and
updates it so building the software will still produce binaries based on
the v1 module until the v2 module is fully released.