database: Introduce BlockSerializer interface.
This introduces a new interface named BlockSerializer and updates the
functions that currently take a pointer to a dcrutil.Block struct to
accept the interface instead.

This removes the tight coupling between the two packages at the API
boundary and allows callers to easily provide custom values.

Since the interface is already implemented by dcrutil.Block, no further
changes to the code or tests is needed.