release: Freeze version 1 blockchain/stake module.
This serves as the final release of version 1 of the blockchain/stake
module.  All future releases will be moving to version 2 of the module.

Consequently, it bumps the required module version as follows:

- github.com/decred/dcrd/blockchain/stake@v1.2.0

It also removes the blockchain/stake override in the root module and
updates it so building the software will still produce binaries based on
the v1 module until the v2 module is fully released.