stake: Accept AddressParams for ticket commit addr.
This modifies the AddrFromSStxPkScrCommitment function to accept a
dcrutil.AddressParams interface instead of a concrete chaincfg.Params to
remove the tight coupling between the two packages at the API boundary
and allow callers to easily provide custom values if desired.  # Please
enter the commit message for your changes. Lines starting