chaincfg: Fix merkle root of v1 simnet genesis block.
The merkle root of the simnet parameters in the v1 API is incorrect,
defined as the same merkle root as of the mainnet genesis block.  This
bug was fixed by chaincfg/v2, making simnets incompatible.  This
change modifies the simnet genesis block of the v1 API to specify the
same (and correct) merkle root as used by the v2 API.

While this is a breaking and hard-forking change, simnet is a private
and disposable network, and fixing this as a patch release allows
simnet compatability between processes using different major versions
of the chaincfg modules.

Testnet 3 has a similar bug, but it is not being patched here due to
testnet being a public network and this being a hard forking change.
chaincfg/v2 did not fix the testnet bug for this reason.