release: Freeze version 3 rpcclient module.
This serves as the final release of version 3 of the rpcclient module.
All future releases will be moving to version 4 of the module.

Since no changes have been made since the last release, it merely
removes the rpcclient override in the root module and updates it so
building the software will still produce binaries based on the v3 module
until the v4 module is fully released.