blockchain: Use standalone module subsidy calcs.
This updates blockchain to make use of the new better optimized subsidy
cache and calculation functions in the the standalone module.  While
making the updates, it also does basic cleanup on some error messages.

Since the v1 chaincfg.Params does not implement the necessary interface
and moving chaincfg/v2.Params requires a major version bump, a new type
named subsidyParams is introduced which embeds v1 chaincfg.Params and
implements the interface to act as a bridge.

Also, since the SubsidyCache type is returned from blockchain, the type
is aliased and rewritten in terms of the standalone module so callers
still on the v1 module are able to interop.

Finally, this also deprecates the related exported functions in the
blockchain module so they can be removed in the next major version.