blockchain/standalone: Prepare v1.0.0.
This updates the documentation for the new blockchain/standalone module
along with main module dependencies to ensure it is tested as well as
uses the latest version of the code as the new module is integrated.

It also serves as a base for blockchain/stake/v1.0.0.

The new modules has the following direct dependencies:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.1
- github.com/decred/dcrd/wire@v1.2.0