release: Freeze version 1 blockchain module use.
This freezes the root module usage of the blockchain module by removing
the replacement and bumping the required version.  This means building
the software will still produce binaries based on the v1 module until
the v2 module is fully released.

All future releases will be moving to version 2 of the module.

Consequently, it bumps the required module versions as follows:

- github.com/decred/dcrd/blockchain@v1.2.0