peer: Accept hash slice for block locators.
This modifies the PushGetBlocksMsg and PushGetHeadersMsg functions to
accept a string of block loactors hashes instead of a concrete
blockchain.BlockLocator to remove the tight couping between the two
packages at the API boundary and allow callers to easily provide custom
values if desired.

It also removes the dependency on the blockchain module since it is no
longer required.