release: Freeze version 2 mempool module use.
This freezes the root module usage of v2 of the mempool module by
removing the replacement and bumping the required version.  This means
building the software will still produce binaries based on the v2 module
until the v3 module is fully released.

All future releases will be moving to version 3 of the module.

Consequently, it bumps the required module versions as follows:

- github.com/decred/dcrd/mempool/v2 v2.1.0