gcs: Add filter version support.
This adds a version parameter to the gcs filters in order to pave the
way for supporting v2 filters which will have a different serialization
that makes them incompatible with v1 filters while still retaining the
ability to work with v1 filters in the interim.

A new error code is added, the tests have been updated accordingly, and
new tests are added to ensure the expected error is thrown on an
unsupported filter version.