main: More complete block logging during sync
In response to issue #1746, enable the logging of the last few
processed blocks, which were previously missed because the logging was
solely dependent on duration. Now by relying on the presently mined
block height, the last few blocks are captured and logged.

Also provide a "Blockchain sync complete" message when all known blocks
are finished processing.