gcs: Support independent fp rate and bin size.
This modifies the code to support an independent false positive rate and
Golomb coding bin size.  Among other things, this permits more optimal
parameters for minimizing the filter size to be specified.

This capability will be used in the upcoming version 2 filters that will
ultimately be included in header commitments.

For a concrete example, the current version 1 filter for block 89341 on
mainnet contains 2470 items resulting in a full serialized size of 6,669
bytes.  In contrast, if the optimal parameters were specified as
described by the comments in this commit, with no other changes to the
items included in the filter, that same filter would be 6,505 bytes,
which is a size reduction of about 2.46%.  This might not seem like a
significant amount, but consider that there is a filter for every block,
so it really adds up.

Since the internal filter no longer directly has a P parameter, this
moves the method to obtain it to the FilterV1 type and adds a new test
to ensure it is returned properly.

Additionally, all of the tests are converted to use the parameters while
retaining the same effective parameters to help prove correctness of the
new code.

Finally, it also significantly reduces the number of allocations
required to construct a filter resulting in faster filter construction
and reduced pressure on the GC and does some other minor consistency
cleanup while here.

In terms of the reduction in allocations, the following is a before and
after comparison of building filters with 50k and 100k elements:

benchmark                    old ns/op    new ns/op     delta
--------------------------------------------------------------
BenchmarkFilterBuild50000    18095111     15680001     -13.35%
BenchmarkFilterBuild100000   31980156     31389892     -1.85%

benchmark                    old allocs   new allocs   delta
--------------------------------------------------------------
BenchmarkFilterBuild50000    31           6            -80.65%
BenchmarkFilterBuild100000   34           6            -82.35%

benchmark                    old bytes    new bytes    delta
--------------------------------------------------------------
BenchmarkFilterBuild50000    1202343      688271       -42.76%
BenchmarkFilterBuild100000   2488472      1360000      -45.35%