gcs: Correct zero hash filter matches.
This ensures filters properly match search items which happen to hash to
zero and adds a test for the condition.  While here, it also rewrites
the MatchAny function to make it easier to reason about.

This was discovered by the new tests which intentionally added tests
with a high false positive rate and random keys.