dcrctl: Expose latest wallet privacy commands.
This updates the wallet rpc/jsonrpc/types module dependency requirement
to the latest version which includes new privacy-related RPCs.

NOTE: This is currently pointing to a fork and will need to updated to
point to a tagged version of the module, once it is available, prior to
be being merged.