chaincfg: Introduce agenda for header cmtmts vote.
This adds a new definition for the upcoming agenda vote to enable header
commitments a provide an initial commitment to a version 2 GCS filter as
defined by DCP0005.  It does not include any code to make decisions or
bump the versions.  It is only the definition.

Also, bump the chaincfg module to v2.3.0 so the new definitions are
available to consumers.