multi: Implement combined merkle root and vote.
This implements the agenda for voting on combining the regular and stake
tree merkle roots into a single merkle root that is stored in the merkle
root field of the block header as defined in DCP0005 along with
consensus tests.

In particular, once the vote has passed and is active, the final merkle
root is the result of a merkle tree that itself has the individual
merkle roots of the two transaction trees as leaves.

It is also worth noting that this does not repurpose the stake root
field once the vote has passed as that will be done in a future commit.
This implies that it is important for both this commit and the
aforementioned future commit which changes the stake root field
semantics to be merged at the same time.

The following is an overview of the changes:

- Generate new version blocks and reject old version blocks after a
  super majority has been reached
  - New block version on mainnet is version 7
  - New block version on testnet is version 8
- Generate block templates with the correct merkle root set in
  accordance with the state of the vote
- Introduce a convenience function for determining if the vote passed
  and is now active
- Introduce a new function in the standalone module for calculating the
  combined merkle root
  - Add tests to ensure the combined merkle root calculation produces
    the correct results
- Modify block validation to enforce the merkle root field in accordance
  with the state of the vote
- Add tests for determining if the agenda is active for both mainnet and
  testnet