gcs/blockcf2: Implement v2 block filter creation.
This implements logic to create version 2 block filters as defined in
DCP0005.

In order to accomplish this a new package named blockcf2 is provided.

Since version 2 block filters require all previous output scripts
referenced as inputs by the block, the new package provides an interface
named PrevScripter to prevent tight coupling to any specific type.  This
provides the caller with the flexibility to provide the necessary
scripts and associated script versions from whatever machanism they deem
fit.

The primary mechanism for creating a block filter with the package is
the Regular function which accepts the block to create the filter for
along with the aforemented PrevScripter interface.

In addition, the package exports the B and M constants which represent
the predefined GCS bin size and false positive rate for the block
filters.  These constants are required both to create the filters and to
deserialize the filters for matching against them.

Finally, the package also provides a Key function which returns the
block-specifc key for the block filter from the merkle root of the
associated block.  The resulting key is required both to create a filter
and to match entries in it.

This only implements the required logic to generate the filters.  Code
to create, store, commit to, validate, and retrieve the filters will be
added in future commits.

A consensus vote is required to commit to the filters and reject blockc
that violate the commitment.  Code to selectively enable consensus
enforcement based on the result of an agenda vote will be added in a
future commit.

The following is a high level overview of the changes:

- Introduce a new package named blockcf2 for creating the filters
- Provide the B and M constants which represent the predefined bin size
  and false positive rate for the block filters
- Introduce PrevScripter interface and associated PrevScriptError
- Implement Regular functions for creating and returning the filter
- Provide Key function to return the block-specific key for a filter
  from the merkle root of the associated block
- Add a README.md for the new package