wire: Implement getcfilterv2/cfilterv2 messages.
This implements new getcfilterv2 and cfilterv2 messages which are used
to request and deliver version 2 committed gcs filters along with their
associated header commitment inclusion proof.

The getcfilterv2 message requests a version 2 gcs filter and proof for a
given block hash.  The cfilterv2 message is sent in response with the
requested filter and proof.

It should be noted that since all header commitments require consensus
changes, the proof can only be used once a consensus vote passes and the
header actually commits to the filter.

Finally, it also bumps the minor default user agent version for the wire
package to account for the updates.